+++ sopa/patches/_transcripts.py
@@ -61,7 +61,7 @@ class OnDiskTranscriptPatches(Patches2D):
 
         self.setup_patches_directory()
 
-        patches_geo_df = gpd.GeoDataFrame(geometry=self.polygons)
+        patches_geo_df = self.geo_df.geometry.to_frame()
 
         if self.write_cells_centroids:
             centroids = self.get_prior_centroids()
@@ -112,7 +112,7 @@ class OnDiskTranscriptPatches(Patches2D):
 
         assert len(valid_indices), "No valid patches found. Check the minimum number of points or cells per patch."
 
-        geo_df = self.as_geodataframe().iloc[valid_indices]
+        geo_df = self.geo_df.iloc[valid_indices]
         geo_df[SopaKeys.CACHE_PATH_KEY] = geo_df.index.map(lambda index: str(self.cache_dir / str(index)))
         geo_df[SopaKeys.POINTS_KEY] = self.points_key